+++ super_editor/lib/src/super_reader/tasks.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import 'package:super_editor/src/core/document.dart';
+import 'package:super_editor/src/default_editor/layout_single_column/layout_single_column.dart';
+import 'package:super_editor/src/default_editor/tasks.dart';
+import 'package:super_editor/src/default_editor/text_tools.dart';
+
+/// Builds [TaskComponentViewModel]s and [TaskComponent]s for every
+/// [TaskNode] in a document.
+///
+/// A [TaskComponent] built by this builder is read-only, meaning that
+/// the user cannot toggle it.
+class ReadOnlyTaskComponentBuilder implements ComponentBuilder {
+  const ReadOnlyTaskComponentBuilder();
+
+  @override
+  TaskComponentViewModel? createViewModel(Document document, DocumentNode node) {
+    if (node is! TaskNode) {
+      return null;
+    }
+
+    final textDirection = getParagraphDirection(node.text.toPlainText());
+
+    return TaskComponentViewModel(
+      nodeId: node.id,
+      padding: EdgeInsets.zero,
+      indent: node.indent,
+      isComplete: node.isComplete,
+      setComplete: null,
+      text: node.text,
+      textDirection: textDirection,
+      textAlignment: textDirection == TextDirection.ltr ? TextAlign.left : TextAlign.right,
+      textStyleBuilder: noStyleBuilder,
+      selectionColor: const Color(0x00000000),
+    );
+  }
+
+  @override
+  Widget? createComponent(
+      SingleColumnDocumentComponentContext componentContext, SingleColumnLayoutComponentViewModel componentViewModel) {
+    if (componentViewModel is! TaskComponentViewModel) {
+      return null;
+    }
+
+    return TaskComponent(
+      key: componentContext.componentKey,
+      viewModel: componentViewModel,
+    );
+  }
+}
